# Only determine view's settings once per lint