--- a/lint/linter.py
+++ b/lint/linter.py
@@ -697,8 +697,7 @@ class Linter(metaclass=LinterMeta):
         Notable: `<path>` can be a list/tuple or str
 
         """
-        settings = self.get_view_settings()
-        executable = settings.get('executable', None)
+        executable = self.settings.get('executable', None)
         if executable:
             persist.debug(
                 "{}: wanted executable is '{}'".format(self.name, executable)
@@ -719,7 +718,7 @@ class Linter(metaclass=LinterMeta):
 
     def insert_args(self, cmd):
         """Insert user arguments into cmd and return the result."""
-        args = self.build_args(self.get_view_settings())
+        args = self.build_args()
         cmd = list(cmd)
 
         if '*' in cmd:
@@ -734,12 +733,9 @@ class Linter(metaclass=LinterMeta):
 
         return cmd
 
-    def get_user_args(self, settings=None):
+    def get_user_args(self):
         """Return any args the user specifies in settings as a list."""
-        if settings is None:
-            settings = self.get_view_settings()
-
-        args = settings.get('args', [])
+        args = self.settings.get('args', [])
 
         if isinstance(args, str):
             args = shlex.split(args)
@@ -748,7 +744,7 @@ class Linter(metaclass=LinterMeta):
 
         return args
 
-    def build_args(self, settings):
+    def build_args(self):
         """
         Return a list of args to add to cls.cmd.
 
@@ -783,16 +779,16 @@ class Linter(metaclass=LinterMeta):
 
         Return the arg list.
         """
-        args = self.get_user_args(settings)
+        args = self.get_user_args()
         args_map = getattr(self, 'args_map', {})
 
         for setting, arg_info in args_map.items():
             prefix = arg_info['prefix']
 
-            if setting not in settings or setting[0] == '@' or prefix is None:
+            if setting not in self.settings or setting[0] == '@' or prefix is None:
                 continue
 
-            values = settings[setting]
+            values = self.settings[setting]
 
             if values is None:
                 continue
@@ -867,10 +863,8 @@ class Linter(metaclass=LinterMeta):
 
         - Add the name/value pair to options.
         """
-        view_settings = self.get_view_settings()
-
         for name, info in self.args_map.items():
-            value = view_settings.get(name)
+            value = self.settings.get(name)
 
             if value:
                 value = util.convert_type(value, type_map.get(name), sep=info.get('sep'))
@@ -881,9 +875,9 @@ class Linter(metaclass=LinterMeta):
 
                     options[name] = value
 
-    def get_working_dir(self, settings):
+    def get_working_dir(self):
         """Return the working dir for this lint."""
-        cwd = settings.get('working_dir', None)
+        cwd = self.settings.get('working_dir', None)
 
         if cwd:
             if os.path.isdir(cwd):
@@ -897,9 +891,9 @@ class Linter(metaclass=LinterMeta):
 
         return self._guess_project_path(self.view.window(), self.view.file_name())
 
-    def get_environment(self, settings):
+    def get_environment(self):
         """Return runtime environment for this lint."""
-        return ChainMap({}, settings.get('env', {}), self.env, BASE_LINT_ENVIRONMENT)
+        return ChainMap({}, self.settings.get('env', {}), self.env, BASE_LINT_ENVIRONMENT)
 
     def get_error_type(self, error, warning):  # noqa:D102
         if error:
@@ -921,6 +915,8 @@ class Linter(metaclass=LinterMeta):
         """
         if self.disabled:
             return []
+
+        self.settings = self.get_view_settings()
 
         # `cmd = None` is a special API signal, that the plugin author
         # implemented its own `run`
@@ -1320,9 +1316,8 @@ class Linter(metaclass=LinterMeta):
         elif not code:
             cmd.append(self.filename)
 
-        settings = self.get_view_settings()
-        cwd = self.get_working_dir(settings)
-        env = self.get_environment(settings)
+        cwd = self.get_working_dir()
+        env = self.get_environment()
 
         if persist.debug_mode():
             util.printf('{}: {} {}'.format(
@@ -1342,9 +1337,8 @@ class Linter(metaclass=LinterMeta):
 
     def tmpfile(self, cmd, code, suffix=''):
         """Run an external executable using a temp file to pass code and return its output."""
-        settings = self.get_view_settings()
-        cwd = self.get_working_dir(settings)
-        env = self.get_environment(settings)
+        cwd = self.get_working_dir()
+        env = self.get_environment()
 
         if persist.debug_mode():
             util.printf('{}: {} {}'.format(