We just call get_view_settings in lint and cache it as an instance variable. Too many methods were passing around or re-determining the settings during a single lint for no apparent reason, since it will not
have changed in the meantime.

Fixes #907.