@kaste 2018-01-30
`self` is not a proper cache for concurrent processing. Passing arguments around is not an 'issue' or problem. 

It is not clear what you're actually trying to solve here. I mean 4 weeks ago we used `self` as a cache for things. and we even had something like `self.code` instead of passing `code` down. 

This is generally an anti-pattern. In that way you usually build wormholes. You put stuff in here and somewhere in a distance you read it again. 

But really, on the first `self.settings = ...` in `lint` we have a race. We cannot do that.