@FichteFoll 2018-01-30
True, it seems like I made a misassumption about when linters are instantiated and used. Notably, I forgot that the `code` to be linted was supplied in an argument to `lint` (and also `hit_time`, but that's irrelevant here) and though that a linter was just `.lint(hit_time)`, in which case caching the settings in an attribute would work.

Seems like we need to pass down the settings in arguments after all. 😒