@kaste 2018-01-30
We would actually pass it down, but we hit exposed API in between like `get_cmd` or `run` and we cannot change the fn signatures here. (That's actually why you called `get_view_settings` as late as possible just before getting into `util` as we developed the `working_dir` stuff.)