@FichteFoll 2018-01-30
Still not too happy about the idea. Let's think about it this way:

When can a view's linter settings change and how frequently? What are the possible race conditions? Under what heuristic can we establish a cache that is loose on the recomputation but responds to underlying changes reasonably as well as not affected by race conditions?