@kaste 2018-01-30
We cannot observe the dependencies of `get_view_settings`, and if you were to write a pure `_get_view_settings` to calculate the arguments to that functions so that it is memoizable takes probably as much time as computing just the result ad-hoc. 

But, if you're talking about efficiency, you should measure the current performance first.  Now, the other problem here is consistency. If we call `get_view_settings` multiple times during one task, bc we cannot pass it down bc of API constraints, we can get different settings at different levels of our code, which would be super-fun to debug of course.

The general solution here would be to use a `threading.local` obj.