@FichteFoll 2018-01-31
>The general solution here would be to use a threading.local obj.

That would work, but only for the settings requests occuring within Linter methods. I have since found out that the majority of settings construction happens in `backend.get_linters` (or that's what it looks like) and am currently optimizing a smart cache around the settings. This is easy for the persisted settings, but I wonder whether I have to poll the window's project_data or whether it is alright to assume that it's only changed when the project file was edited within ST itself. I don't really think so, though, so I'm gonna set up some monitoring for that.