@FichteFoll 2018-01-31
All right, I think this is it. Now settings are only re-evaluated when settings or variables (i.e. folders or file name) actually change. The settings check is thread-safe through a mutex (didn't see any other way) but stays consistent along the execution of a single `.lint` call. PTAL.