@kaste 2018-01-31
I'm totally unhappy here. The sheer amount of work you did here, makes it harder to reject this PR. But I think we should seek for a simpler solution here. First, if we want to cache something, we need to spell out all its dependencies as I said in a comment above. Here, we still have the possibility to serve stale settings. 

Now, the window settings object is a cache around `window.project_data()` essentially, but we cannot have a performance problem for such a call. Honestly, I don't think we have a performance problem at all here.

We have a consistency problem here, maybe, actually really unlikely. This could be tackled. But a solution to this must/should also state in the code what we're trying to solve.