@braver 2018-01-31
> Honestly, I don't think we have a performance problem at all here.

I don’t know the inner workings of what we do with settings, so I may be completely off here. I do know that reading settings via Sublime’s api is super cheap, it basically immediately returns the value from memory. Settings shouldn’t be cached and reading them shouldn’t be optimized. Now perhaps we do processing on the values, since we don’t have simple key-value pairs, that’s expensive... I don’t know. I really really don’t want to introduce a cache invalidation problem, certainly not if it’s about reading settings, and also not if we haven’t proven a performance issue and measured the performance improvement of this proposal.