@FichteFoll 2018-02-01
## Testing procedure

1. Apply patch to measure performance to various branches (see below).
1. Disable debug mode.
1. Restart ST as is, with 4 windows and 39 views.
1. Upon restarting, ST opens with the sublimelinter window in focus, having two panes and the files `SublimeLinter/lint/backend.py` and `SublimeLinter-contrib-mypy/linter.py` active (not the SL linter one).
1. Open the console to observe the time it took to lint each file.
1. Repeat multiple times from 3 to get an idea of the average.

Patch to measure performance:

```diff
diff --git a/lint/backend.py b/lint/backend.py
index 0b55738..427328f 100644
--- a/lint/backend.py
+++ b/lint/backend.py
@@ -32,6 +32,8 @@ def lint_view(view, hit_time, callback):
     aggregated, and for each selector, if it occurs in sections,
     the corresponding section is linted as embedded code.
     """
+    import time
+    start = time.time()
     linters = get_linters(view)
     lint_tasks = get_lint_tasks(linters, view)
 
@@ -40,6 +42,9 @@ def lint_view(view, hit_time, callback):
         for task in lint_tasks)
 
     all_errors = chain.from_iterable(results)
+    all_errors = list(all_errors)
+    end = time.time()
+    print("linting {!r} took {:.3f}s".format(view.file_name(), end - start))
 
     # We don't want to guarantee that our consumers/views are thread aware.
     # So we merge here into Sublime's shared worker thread. Sublime guarantees
```

---

## Results

| both PRs | `linter.py` | `backend.py`
|-:|-:|-:|
| *1* | 1.972 | 3.242 |
| *2* | 2.646 | 2.655 |
| *3* | 2.075 | 3.060 |
| *4* | 2.024 | 3.579 |
| *5* | 2.218 | 3.280 |


| this PR | `linter.py` | `backend.py`
|-:|-:|-:|
| *1* | 3.749 | 3.868 |
| *2* | 3.574 | 4.062 |
| *3* | 2.311 | 2.935 |
| *4* | 3.237 | 3.642 |
| *5* | 2.704 | 3.272 |


| master\* | `linter.py` | `backend.py`
|-:|-:|-:|
| *1* | 3.825 | 4.178 |
| *2* | 3.677 | 3.634 |
| *3* | 3.609 | 3.854 |
| *4* | 4.313 | 4.337 |
| *5* | 4.489 | 4.868 |


| Averages | `linter.py` | `backend.py` | both | % |
|:-|-:|-:|-:|-:|
| both PRs | 2.187  | 3.1628 | 2.6749 | 100% |
| this PR  | 3.115  | 3.5558 | 3.3354 | 124% |
| master\* | 3.9826 | 4.1742 | 4.0784 | 152% |

\* dde041249590c7f0f56916e441fa09738897d0fa (essentially `master` without #904)